server/requirements.txt
appdirs==1.4.4
astroid==2.4.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dotenv==0.14.0
six==1.14.0
toml==0.10.0
urllib3==1.25.9
virtualenv==20.0.23
Werkzeug==1.0.1
wrapt==1.12.1
